exported_scraping_work/requirements.txt
google-cloud-bigquery>=3.0.0
google-cloud-aiplatform>=1.0.0
google-cloud-logging>=3.0.0
praw>=7.7.0
pyahocorasick>=2.0.0
requests>=2.28.0
functions-framework>=3.0.0
